* libsvn_client/commit_util.c: Update file name in comment at top of file. (svn_client__do_commit): Silence signed/unsigned warning on Windows: make common_len a size_t, which is what strlen should be returning.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@842146 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/commit_util.c
@@ -1,5 +1,5 @@
 /*
- * commit-drive.c:  Driver for the WC commit process.
+ * commit_util.c:  Driver for the WC commit process.
  *
  * ====================================================================
  * Copyright (c) 2000-2002 CollabNet.  All rights reserved.
@@ -917,7 +917,7 @@ svn_client__do_commit (const char *base_url,
     {
       const char *item_url, *item_dir, *item_name;
       const char *common = NULL;
-      int common_len;
+      size_t common_len;
       svn_client_commit_item_t *item
         = ((svn_client_commit_item_t **) commit_items->elts)[i];
